requirements.txt
aiogram==2.19
python-dotenv==0.21.0
gino==1.0.1
SQLAlchemy~=1.3.24
asyncpg~=0.26.0